+++ examples/groq/smart_scraper_groq_openai.py
@@ -1,36 +1,43 @@
 """ 
-Basic example of scraping pipeline using ScriptCreatorGraph
+Basic example of scraping pipeline using SmartScraper
 """
 
 import os
 from dotenv import load_dotenv
-from scrapegraphai.graphs import ScriptCreatorGraph
+from scrapegraphai.graphs import SmartScraperGraph
 from scrapegraphai.utils import prettify_exec_info
 
 load_dotenv()
+
 
 # ************************************************
 # Define the configuration for the graph
 # ************************************************
 
+groq_key = os.getenv("GROQ_APIKEY")
 openai_key = os.getenv("OPENAI_APIKEY")
 
 graph_config = {
     "llm": {
+        "model": "groq/gemma-7b-it",
+        "api_key": groq_key,
+        "temperature": 0
+    },
+    "embeddings": {
         "api_key": openai_key,
         "model": "gpt-3.5-turbo",
     },
-    "library": "beautifulsoup"
+    "headless": False
 }
 
 # ************************************************
-# Create the ScriptCreatorGraph instance and run it
+# Create the SmartScraperGraph instance and run it
 # ************************************************
 
-smart_scraper_graph = ScriptCreatorGraph(
-    prompt="List me all the news with their description.",
+smart_scraper_graph = SmartScraperGraph(
+    prompt="List me all the projects with their description.",
     # also accepts a string with the already downloaded HTML code
-    source="https://perinim.github.io/projects",
+    source="https://perinim.github.io/projects/",
     config=graph_config
 )