+++ scrapegraphai/helpers/robots.py
@@ -1,7 +1,7 @@
-
 """ 
 Module for mapping the models in ai agents
 """
+
 robots_dictionary = {
     "gpt-3.5-turbo": ["GPTBot", "ChatGPT-user"],
     "gpt-4-turbo": ["GPTBot", "ChatGPT-user"],