+++ scrapegraphai/nodes/fetch_node.py
@@ -1,5 +1,5 @@
 """ 
-Module for fetching the HTML node
+FetchNode Module
 """
 
 from typing import List, Optional
@@ -12,38 +12,24 @@ from ..utils.remover import remover
 class FetchNode(BaseNode):
     """
     A node responsible for fetching the HTML content of a specified URL and updating
-    the graph's state with this content. It uses the AsyncHtmlLoader for asynchronous
-    document loading.
+    the graph's state with this content. It uses the AsyncChromiumLoader to fetch the
+    content asynchronously.
 
     This node acts as a starting point in many scraping workflows, preparing the state
     with the necessary HTML content for further processing by subsequent nodes in the graph.
 
     Attributes:
-        node_name (str): The unique identifier name for the node.
-        node_type (str): The type of the node, defaulting to "node". This categorization
-                         helps in determining the node's role and behavior within the graph.
-                         The "node" type is used for standard operational nodes.
-
+        headless (bool): A flag indicating whether the browser should run in headless mode.
+        verbose (bool): A flag indicating whether to print verbose output during execution.
+    
     Args:
-        node_name (str): The unique identifier name for the node. This name is used to
-                         reference the node within the graph.
-        node_type (str, optional): The type of the node, limited to "node" or
-                                   "conditional_node". Defaults to "node".
-
-    Methods:
-        execute(state): Fetches the HTML content for the URL specified in the state and
-                        updates the state with this content under the 'document' key.
-                        The 'url' key must be present in the state for the operation
-                        to succeed.
+        input (str): Boolean expression defining the input keys needed from the state.
+        output (List[str]): List of output keys to be updated in the state.
+        node_config (Optional[dict]): Additional configuration for the node.
+        node_name (str): The unique identifier name for the node, defaulting to "Fetch".
     """
 
     def __init__(self, input: str, output: List[str], node_config: Optional[dict], node_name: str = "Fetch"):
-        """
-        Initializes the FetchHTMLNode with a node name and node type.
-        Arguments:
-            node_name (str): name of the node
-            prox_rotation (bool): if you wamt to rotate proxies
-        """
         super().__init__(node_name, "node", input, output, 1)
 
         self.headless = True if node_config is None else node_config.get("headless", True)
@@ -55,13 +41,14 @@ class FetchNode(BaseNode):
         update the state with this content.
 
         Args:
-            state (dict): The current state of the graph, expected to contain a 'url' key.
+            state (dict): The current state of the graph. The input keys will be used
+                            to fetch the correct data types from the state.
 
         Returns:
-            dict: The updated state with a new 'document' key containing the fetched HTML content.
+            dict: The updated state with a new output key containing the fetched HTML content.
 
         Raises:
-            KeyError: If the 'url' key is not found in the state, indicating that the
+            KeyError: If the input key is not found in the state, indicating that the
                     necessary information to perform the operation is missing.
         """
         if self.verbose: