+++ scrapegraphai/utils/token_calculator.py
@@ -8,15 +8,21 @@ from ..helpers.models_tokens import models_tokens
 
 def truncate_text_tokens(text: str, model: str, encoding_name: str) -> List[str]:
     """
-    It creates a list of strings to create max dimension tokenizable elements
+    Truncates text into chunks that are small enough to be processed by specified llm models.
 
     Args:
-        text (str): The input text to be truncated into tokenizable elements.
-        model (str): The name of the language model to be used.
-        encoding_name (str): The name of the encoding to be used (default: EMBEDDING_ENCODING).
+        text (str): The input text to be truncated.
+        model (str): The name of the llm model to determine the maximum token limit.
+        encoding_name (str): The encoding strategy used to encode the text before truncation.
 
     Returns:
-        List[str]: A list of tokenizable elements created from the input text.
+        List[str]: A list of text chunks, each within the token limit of the specified model.
+
+    Example:
+        >>> truncate_text_tokens("This is a sample text for truncation.", "GPT-3", "EMBEDDING_ENCODING")
+        ["This is a sample text", "for truncation."]
+
+    This function ensures that each chunk of text can be tokenized by the specified model without exceeding the model's token limit.
     """
 
     encoding = tiktoken.get_encoding(encoding_name)